Change Vector2 x y to float

+++ dev/shirako/reain/Vector2.java
@@ -1,10 +1,10 @@
 package dev.shirako.reain;
 
 public class Vector2 {
-    public int x;
-    public int y;
+    public float x;
+    public float y;
 
-    public Vector2(int x, int y) {
+    public Vector2(float x, float y) {
         this.x = x;
         this.y = y;
     }
